added viewModel
